Fix indentation and syntax in js files.
This commit is not exhaustive and more work in that sense could be achieved.

- In .jshintrc, `indent: 2` is set. In many files, we have a mix of 2 and 4 spaces-indented blocks.
This commit tries to fix most of the files with a mix of different indentation. Some files with 4 spaces indented code have been left as is.
- I change syntax from `function *myGenerator` to `function* myGenerator` to follow https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/function*

This commit is tedious to review but the code lacks consistency in style and yet tools have been set up. This could enforced more strongly to avoid further divergence in style.

"Auto Linter on save" could be a solution if everyone agree on the tool. Hope that helps, cheers